ctags: add back var indexing for go files
It was disabled previously since it used to pick up bindata vars which
slowed down :tag lookup in vim. This has been fixed by switch ctags to
generate "tag" files which record the line number instead of a pattern.

Since gutentags will keep the tag file updates on changes, the switch
from using regular expressions to jump to the tag in it's file to simply
jumping to a line number should be large unnoticable.